# api/docs/swagger.yaml
definitions:
...
  models_v1.ChangePassword:
    properties:
      code:
        example: "111111"
        type: string
      password:
        example: password_1
        type: string
      source:
        example: example@example.com
        type: string
      type:
        enum:
        - email
        - phone_number
        example: email
        type: string
    required:
    - code
    - password
    - source
    - type
    type: object
...
  models_v1.LoginRequest:
    properties:
      login:
        example: example@example.com
        type: string
      password:
        example: password_1
        type: string
    required:
    - login
    - password
    type: object
  models_v1.RegisterRequest:
    properties:
      email:
        example: example@example.com
        type: string
      name:
        example: Islombek
        type: string
      password:
        example: password_1
        type: string
      phone_number:
        example: "998912345678"
        type: string
    required:
    - email
    - name
    - password
    - phone_number
    type: object
  models_v1.RequestCode:
    properties:
      email:
        type: string
      need_code:
        type: boolean
      phone_number:
        type: string
    type: object
  models_v1.RequestCodeRequest:
    properties:
      source:
        example: example@example.com
        type: string
      type:
        enum:
        - email
        - phone_number
        example: email
        type: string
    required:
    - source
    - type
    type: object
  models_v1.Response:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  models_v1.Token:
    properties:
      token:
        type: string
    type: object
...
  models_v1.VerifyCodeRequest:
    properties:
      code:
        example: "111111"
        type: string
      source:
        example: example@example.com
        type: string
      type:
        enum:
        - email
        - phone_number
        example: email
        type: string
    required:
    - code
    - source
    - type
    type: object
host: localhost:8123
info:
  contact: {}
  title: E-commerce project v2
paths:
...
  /api/auth/change_password:
    post:
      consumes:
      - application/json
      description: Change password works only for clients.
      operationId: changePassword
      parameters:
      - description: Change password
        in: body
        name: change_password
        required: true
        schema:
          $ref: '#/definitions/models_v1.ChangePassword'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully changed password
          schema:
            $ref: '#/definitions/models_v1.Response'
        "400":
          description: Bad Request, invalid password
          schema:
            $ref: '#/definitions/models_v1.Response'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/models_v1.Response'
        "406":
          description: Invalid verification code
          schema:
            $ref: '#/definitions/models_v1.Response'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/models_v1.Response'
        "501":
          description: Not implemented"s
          schema:
            $ref: '#/definitions/models_v1.Response'
      summary: ChangePassword
      tags:
      - auth
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Login
      operationId: login
      parameters:
      - description: Login request
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/models_v1.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success returning token
          schema:
            $ref: '#/definitions/models_v1.Token'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models_v1.Response'
        "406":
          description: User not verified
          schema:
            $ref: '#/definitions/models_v1.Response'
        "417":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/models_v1.Response'
        "500":
          description: Internal
          schema:
            $ref: '#/definitions/models_v1.Response'
      summary: Login
      tags:
      - auth
  /api/auth/login_admin:
    post:
      consumes:
      - application/json
      description: Login for admin. This is need to enter for admin panel
      operationId: loginadmin
      parameters:
      - description: Login request
        in: body
        name: login_admin
        required: true
        schema:
          $ref: '#/definitions/models_v1.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success returning token
          schema:
            $ref: '#/definitions/models_v1.Token'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models_v1.Response'
        "406":
          description: User not verified
          schema:
            $ref: '#/definitions/models_v1.Response'
        "417":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/models_v1.Response'
        "500":
          description: Internal
          schema:
            $ref: '#/definitions/models_v1.Response'
      summary: Login for admin
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Create Client
      operationId: register
      parameters:
      - description: Register Request
        in: body
        name: register
        required: true
        schema:
          $ref: '#/definitions/models_v1.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success Request that needs verification
          schema:
            $ref: '#/definitions/models_v1.RequestCode'
        "400":
          description: Bad Request / Invalid email / Invalid phone / Invalid password
          schema:
            $ref: '#/definitions/models_v1.Response'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/models_v1.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models_v1.Response'
      summary: Create Client
      tags:
      - auth
  /api/auth/request_code:
    post:
      consumes:
      - application/json
      description: request code is needed when password is forgotten
      operationId: requestCode
      parameters:
      - description: Request code request
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/models_v1.RequestCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success returning token
          schema:
            $ref: '#/definitions/models_v1.Response'
        "400":
          description: Bad request / Invalid email / Invalid phone
          schema:
            $ref: '#/definitions/models_v1.Response'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/models_v1.Response'
        "500":
          description: Internal / Failed to send code
          schema:
            $ref: '#/definitions/models_v1.Response'
        "501":
          description: Not implemented email/phone verification
          schema:
            $ref: '#/definitions/models_v1.Response'
      summary: Request code
      tags:
      - auth
  /api/auth/verify_code:
    post:
      consumes:
      - application/json
      description: Verify client's email/phone
      operationId: verifyCode
      parameters:
      - description: Verify code Request
        in: body
        name: verifyCode
        required: true
        schema:
          $ref: '#/definitions/models_v1.VerifyCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/models_v1.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models_v1.Response'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/models_v1.Response'
        "406":
          description: Invalid verification code
          schema:
            $ref: '#/definitions/models_v1.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models_v1.Response'
        "501":
          description: Not implemented email/phone verification
          schema:
            $ref: '#/definitions/models_v1.Response'
      summary: Verify client's email/phone
      tags:
      - auth
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"